Fix release review findings
